💎 FINAL FIX: Synchronized all UI and Utils with Long IDs and fixed Streaming Service icons

+++ android-app/app/src/main/java/com/duallive/app/data/entity/Standing.kt
@@ -1,12 +1,12 @@
 package com.duallive.app.data.entity
 
 data class Standing(
-    val teamId: Int,
+    val teamId: Long,
     var matchesPlayed: Int = 0,
     var wins: Int = 0,
     var draws: Int = 0,
     var losses: Int = 0,
-    var goalsFor: Int = 0,     // ADDED
-    var goalsAgainst: Int = 0, // ADDED
+    var goalsFor: Int = 0,
+    var goalsAgainst: Int = 0,
     var points: Int = 0
 )
